+++ spec/icon_spec.rb
@@ -8,10 +8,28 @@ describe GameIcons::Icon do
     expect(i.string).to eq('<svg><path fill="abc"/><path fill="def"/></svg>')
   end
 
-  it 'recolors properly' do
+  it 'recolors regular colors' do
+    exp = <<-EOSVG
+<?xml version="1.0"?>
+<svg>
+  <path fill="123" fill-opacity="1.0"/>
+  <path fill="456" fill-opacity="1.0"/>
+</svg>
+EOSVG
     i = GameIcons::Icon.new(data('foo.svg'))
-    expect(i.recolor(bg: '123', fg: '456').string)
-      .to eq("<?xml version=\"1.0\"?>\n<svg>\n  <path fill=\"123\"/>\n  <path fill=\"456\"/>\n</svg>\n")
+    expect(i.recolor(bg: '123', fg: '456').string).to eq(exp)
+  end
+
+  it 'recolors with fill opacity too' do
+    exp = <<-EOSVG
+<?xml version="1.0"?>
+<svg>
+  <path fill="123" fill-opacity="0.25"/>
+  <path fill="456" fill-opacity="0.12"/>
+</svg>
+EOSVG
+    i = GameIcons::Icon.new(data('foo.svg'))
+    expect(i.recolor(bg: '123', fg: '456', bg_opacity:'0.25', fg_opacity:'0.12').string).to eq(exp)
   end
 
   it 'returns the file given' do